add redeploy script

+++ redeploy.sh
@@ -1,6 +1,18 @@
-#!bin/bash
+#!/bin/sh
 
 git checkout develop
-git pull
-docker-compose down
-docker-compose up --build -d
+
+UPSTREAM=${1:-'@{u}'}
+LOCAL=$(git rev-parse @)
+REMOTE=$(git rev-parse "$UPSTREAM")
+
+if [ $LOCAL = $REMOTE ]; then
+    echo $(date) "Up-to-date"
+elif [ $LOCAL = $BASE ]; then
+    echo $(date) "Need to pull"
+    git pull
+    docker-compose down
+    docker-compose up --build -d
+else
+    echo $(date) "Diverged"
+fi
